refactor: enhance availability block meeting types management
- Introduced a new hook for updating meeting types associated with availability blocks.
- Updated the AvailabilityModal to utilize the new hook for better state management.
- Refactored API endpoints to streamline meeting type retrieval and updates.
- Removed the deprecated utility functions related to meeting types.
