+++ src/types/Account.ts
@@ -103,7 +103,7 @@ export interface TimeRange {
 
 export interface AccountPreferences {
   id?: string
-  timezone: string
+  timezone: string | undefined
   availableTypes: MeetingType[]
   description?: string
   availabilities: DayAvailability[]
